feat:提供普通消息传递公用类
